Added java output code
Added a method for each language to add a file header (e.g. all code you
need for a custom JComponent). These methods should read in template
textfiles that contain typical component code for each language and add
it to the exported file. Modified output of the Java2D code.