Modified the conversion to JavaFX 2.0 component and added package info
Now the converter creates a more standardized JavaFX 2.0 custom control with a Control, a Skin, a Behavior and a css file. The css file is not used for styling but with this template it should be easier to create custom JavaFX 2.0 controls.
In addition the converter ui now has a textfield where you could put the package information for the created classes.