Better code style in java and javafx output
Modified the output of Java and JavaFX code to avoid too long lines. Now shapes like Rectangles, Ellipses, Circles etc. will contain line breaks in their definition. Also Linear- and RadialGradients will contain line breaks in their definition.
The fill of a shape will now be stored in an additional variable instead of direct applying it to the shape it belongs to.
So the created code should be better to read now…i hope… :)