Added export of JavaFX fills to css file
Now every shape fill (color, linear- and radial-gradient) will be also exported to the css file when translating to JavaFX. In the code a comment will be added that one could decomment to use the setId("STYLE_ID") method instead of the drawing code. In the future the drawing code might be completely replaced by the css approach.