Modified the export of JavaFX drop shadow and inner shadow
Because the parameters for the drop shadow and inner shadow that are used in Adobe Fireworks are different from the parameters in JavaFX i have made some adjustments that now leads to better results. In addition i've changed the code for the inner shadow that will be created by the converter. Now effect-chaining should also work correctly.