Added support for custom properties in JavaFX output
Now it is possible to create ObjectProperties like ObjectProperty<YOUR_TYPE> with their getters and setters and also with the appropriate changeListeners in the JavaFX skin class of the control.
Supported types in the fxg file are:
- boolean
- double
- int
- long
- string
- object

and all other will be interpreted as custom types e.g.
- LedColor_color
will lead to a ObjectProperty<LedColor> color in the code.