Modified the HTML5 output
Now the controls that will be created are taking parameters. One of these parameters could be an id of a div tag. The control will be injected into tag of the given id, if it is not given, it will inject the canvas tag in the body. In addition i've also added a repaint method that clears all buffers before it redraws.